reorganizing parser files and headerfile
