+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: tsoares- <tsoares-@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/18 08:11:33 by tsoares-          #+#    #+#             */
-/*   Updated: 2024/08/24 23:10:32 by jgils            ###   ########.fr       */
+/*   Updated: 2024/08/25 01:43:44 by jgils            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -62,7 +62,7 @@ int	is_quote(char c);
 int	is_space(char c);
 int	is_metachar(char c);
 
-// PR.C ------------------------------------------------------------
+// COMMAND_TABLE.C ------------------------------------------------------------
 
 
 #endif
